third_party/blink/web_tests/fast/events/touch/gesture/gesture-tap-active-state-hidden-iframe-expected.txt

Gestures go here

Tests that tap gesture events on a an element covered by a hidden iframe set and clear the active state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Verify active isn't initially set
PASS getHoverActiveState(box) is "default"
Verify showPress, tap sets and clears active
PASS getHoverActiveState(box) is "hoveredAndActive"
PASS getHoverActiveState(box) is "hovered"
PASS successfullyParsed is true

TEST COMPLETE
